Fix #14: iostuff.f: 'hbonde' is used uninitialized
Add hbonde to the /simdat/ common block for shared use by subroutines
wrtsim and rddet. This fixes the immediate issue of hbonde being used
uninitialized, but there still remains the issue of hbonde not being
used except for read and write. However, the alternative of removing
hbonde altogether would result in changing the I/O format and breaking
compatibility.